python-client/requirements.txt
cassandra-driver>=3.24.0
pyspark>=3.0.1
pandas>=1.2.0